requirements.txt
aioredis==1.3.1
alembic==1.7.6
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-autoslug==1.9.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-environ==0.9.0
django-filter==21.1
django-phonenumber-field==6.1.0
django-qr-code==3.1.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
email-validator==1.1.3
fcm-django==1.0.12
firebase-admin==5.2.0
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.48.0
grpcio-status==1.48.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
incremental==21.3.0
itsdangerous==2.1.0
Jinja2==3.0.3
jws==0.1.3
kombu==5.2.4
ldap3==2.9.1
Mako==1.1.6
MarkupSafe==2.1.0
msgpack==1.0.3
oauth2client==4.1.3
paramiko==2.9.2
passlib==1.7.2
phonenumbers==8.12.44
Pillow==9.0.1
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
segno==1.5.2
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.31
sqlparse==0.4.2
sshtunnel==0.4.0
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
vonage==2.7.0
wcwidth==0.2.5
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.14.1
WTForms==2.2.1
zipp==3.7.0
zope.interface==5.4.0
